Hemtt Fixes
