+++ addons/hidden_units/CfgVehicles_Baseclasses.hpp
@@ -39,14 +39,11 @@ class CUP_B_CZ_Soldier_DST_Base;
 class CUP_B_CZ_Soldier_WDL_Base;
 class CUP_B_CZ_SpecOp_DST_Base;
 class CUP_B_CZ_SpecOp_WDL_Base;
-class CUP_B_GER_Fleck_Crye2;
-class CUP_B_GER_Fleck_Crye;
 class CUP_B_GER_Soldier_01;
 class CUP_B_GER_Soldier_02;
 class CUP_B_GER_Soldier_03;
 class CUP_B_GER_Soldier_04;
 class CUP_B_GER_Soldier_07;
-class CUP_B_GER_Soldier_Fleck_Overalls_Tank;
 class CUP_B_GER_Soldier_Overalls_Pilot;
 class CUP_B_GER_Soldier_Overalls_Tank;
 class CUP_B_USArmy_OCP_Soldier_01;
