--- addons/hlc_tweaks/hidden_augRifles.hpp
@@ -1,91 +1,0 @@
-// Some cause pop-up errors.
-class hlc_rifle_aug: hlc_aug_base {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_auga1_t: hlc_rifle_aug {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_auga1_b: hlc_rifle_aug {
-    scope = 1;
-    scopeArsenal = 1;
-};
-
-class hlc_rifle_auga1carb: hlc_rifle_aug {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_auga1carb_B: hlc_rifle_auga1carb {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_auga1carb_t: hlc_rifle_auga1carb {
-    scope = 1;
-    scopeArsenal = 1;
-};
-
-class hlc_rifle_augsr: hlc_rifle_auga2 {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augsr_t: hlc_rifle_augsr {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augsr_b: hlc_rifle_augsr {
-    scope = 1;
-    scopeArsenal = 1;
-};
-
-class hlc_rifle_augsrcarb: hlc_rifle_auga2carb {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augsrcarb_t: hlc_rifle_augsrcarb {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augsrcarb_b: hlc_rifle_augsrcarb {
-    scope = 1;
-    scopeArsenal = 1;
-};
-
-class hlc_rifle_augsrhbar: hlc_rifle_auga2lsw {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augsrhbar_t: hlc_rifle_augsrhbar {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augsrhbar_b: hlc_rifle_augsrhbar {
-    scope = 1;
-    scopeArsenal = 1;
-};
-
-class hlc_rifle_augpara: hlc_aug_base {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augpara_b: hlc_rifle_augpara {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_augpara_t: hlc_rifle_augpara {
-    scope = 1;
-    scopeArsenal = 1;
-};
-
-class hlc_rifle_auga2para: hlc_aug_base {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_auga2para_t: hlc_rifle_auga2para {
-    scope = 1;
-    scopeArsenal = 1;
-};
-class hlc_rifle_auga2para_b: hlc_rifle_auga2para {
-    scope = 1;
-    scopeArsenal = 1;
-};
