+++ addons/recoils/niarms/weapons/baseclasses.hpp
@@ -2,8 +2,6 @@ class hlc_acr556_base;
 class hlc_acr_base;
 class hlc_ar15_base;
 class hlc_aug_base;
-class hlc_lmg_minimi_railed;
-class hlc_lmg_minimipara;
 class hlc_rifle_ACR68_SBR_black;
 class hlc_rifle_ACR_SBR_black;
 class hlc_rifle_auga3;
@@ -11,4 +9,3 @@ class hlc_rifle_CQBR;
 class hlc_rifle_honeybase;
 class hlc_saw_base;
 class hlc_sg550_base;
-class hlc_rifle_Bushmaster300;
